Reformat 3.1RC into table / add BitTorrent download
This adds the BitTorrent magnet infohash as a download option.  Of course, verification remains exactly the same as is necessary.  Importantly, Github Markdown doesn't seem to render links to magnet: URIs, but I have chosen to preserve the URI up until the moment they decide to get their shit together and make it happen.

I will be hosting the torrent download indefinitely.  Please double-check: download it using a torrent client and check that the download is valid.