Update link to source code
There's no point in having a page on our website about the source code
if we don't link to it, so we should either consistently link to it
(whenever linking to "Qubes OS source code") or simply remove the page.
I would argue that we should keep the page (and consistently link to it)
rather than linking directly to GitHub since Qubes OS source code is
independent of GitHub. (For example, we could one day decide to switch
from GitHub to a different Git hosting service.)