Add a link to the official page to the "unverified translation" warning
The "unverified translation" warning message could cause the need to go
to the official page. The problem here is that the visitor has to know
how to do that (select `en` (and nothing else)). This could be easier:
The warning message shows a link to the official version of the page.