Improve language switcher position across screen sizes
On large screens, the language switcher will be at the top right of the
page, after the last button in the nav bar ("Donate"). On small screens,
the language switcher will be immediately to the left of the nav menu
collapse toggle (hamburger menu). This commit also improves the font and
spacing of the language items within the menu.